[MIOpen] Fix HIPRTC ambiguous __habs/hsqrt/__hisnan with half_t (TheRock #4852)
Newer HIP headers add __hip_bfloat16 overloads for __habs, hsqrt, and __hisnan
in addition to the existing __half overloads. Because composable_kernel's
half_t is defined as _Float16 (data_type.hpp), calls like __habs(a) where 'a'
is half_t become ambiguous: both __half and __hip_bfloat16 accept implicit
conversion from _Float16.

This breaks JIT (HIPRTC) compilation of MIOpen reduction kernels invoked by
PyTorch on gfx950, gfx120x, and gfx110x, causing test failures such as
test_variable_sequence_cuda_float16/float32, test_autocast_rnn,
test_cudnn_weight_format, and test_rnn_retain_variables_*.

Fix: explicitly cast the operand to __half at every affected call site so
overload resolution unambiguously picks the fp16 intrinsic.

Affected files (all in legacy_composable_kernel):
  - reduction_operator.hpp:    __habs (x2), hsqrt (x1)
  - reduction_functions_binop.hpp: __hisnan (x1)

Verified by recompiling the headers with hipcc --offload-arch=gfx1151;
the four ambiguity errors are gone.

+++ projects/miopen/src/legacy_composable_kernel/composable_kernel/include/utility/reduction_functions_binop.hpp
@@ -34,7 +34,9 @@
 namespace ck {
 namespace detail {
 
-static inline __device__ bool isnan(half_t x) { return __hisnan(x); };
+// Cast to __half to disambiguate __hisnan (also overloaded for __hip_bfloat16
+// in newer HIP headers, which makes the call ambiguous when half_t == _Float16).
+static inline __device__ bool isnan(half_t x) { return __hisnan(static_cast<__half>(x)); };
 
 template <NanPropagation_t nanPropaOpt, typename opReduce, typename compType>
 struct binop_with_nan_check;
